backend/requirements.txt
fastapi
uvicorn
motor == 2.4.0
pyJWT==1.7.1
python-decouple==3.3
pydantic[email]